pylint corrections
